// lib/api/generated/models/pageObjectOfUserAuthResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { PageableObject } from './pageableObject';
import type { SortObject } from './sortObject';
import type { UserAuthResponse } from './userAuthResponse';

export interface PageObjectOfUserAuthResponse {
  content?: UserAuthResponse[];
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}